Extract solang
